+++ zhc_ir/src/annotation/val_use.rs
@@ -6,5 +6,5 @@ pub struct AnnValUseRef<'s, 'ann, D: Dialect, OpAnn: Annotation, ValAnn: Annotat
     /// Annotated reference to the operation that consumes this value.
     pub opref: AnnOpRef<'s, 'ann, D, OpAnn, ValAnn>,
     /// Zero-based index into the consuming operation's argument list.
-    pub position: u8,
+    pub position: u16,
 }
